feat: integrates QuantityInput and RemoveIngredientButton into rest of the components
Creates a new data structure for ingredients, storing the API serving size and nutrition info in the same object as the user quantity and nutrition info. Edits the remove reducer in meal-slice and adds a new to update user quantity and nutrition. Changes arguments and variables
for calculateWeightedValues in addition to exporting the method. Makes a number of minor changes to Row, Total, and Table to integrate the new components and data structure.