feat: turns BarChart into a stacked bar chart
Changes the data structure of chartData and adds the VictoryStack component. On Table, tbodyElement is now assigned to the useMemo hook to prevent the hook from not running when App re-renders (e.g. showing or hiding BarChart). Also, keys were added to the  elements lists on Table and Row to try to fix the previously mentioned bug